infrastructure/dgtocean/automation-commands/droplet-files/requirements.txt
py4j>=0.10.9.7
pandas>=1.0.5
pyarrow>=4.0.0,<13.0.0
numpy>=1.15
grpcio>=1.48,<1.57
grpcio-status>=1.48,<1.57
googleapis-common-protos==1.56.4
pyspark==3.5.3
psycopg2-binary==2.9.9
